Add AI Peer Review Module
