+++ src/ai_peer_review/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AiPeerReviewConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "ai_peer_review"
