+++ src/research_ai/constants.py
@@ -9,7 +9,7 @@ BASE_FRONTEND_URL = getattr(
 
 
 class ExpertiseLevel(models.TextChoices):
-    """Career stage for expert recommendations. Value: snake_case (API/DB), label: display."""
+    """Expert recommendations career stage. Value: snake_case (API/DB), label: display."""
 
     PHD_POSTDOCS = "phd_postdocs", "PhD/PostDocs"
     EARLY_CAREER = "early_career", "Early Career Researchers"
@@ -59,22 +59,6 @@ class EmailTemplateType(models.TextChoices):
 
 VALID_EMAIL_TEMPLATE_KEYS = frozenset(e.value for e in EmailTemplateType)
 DEFAULT_EMAIL_TEMPLATE_KEY = EmailTemplateType.COLLABORATION.value
-
-class ReviewStatus(models.TextChoices):
-    """Async AI job lifecycle for proposal review and RFP summary."""
-
-    PENDING = "pending", "pending"
-    PROCESSING = "processing", "processing"
-    COMPLETED = "completed", "completed"
-    FAILED = "failed", "failed"
-
-
-class OverallRating(models.TextChoices):
-    """Aggregate proposal quality from five dimension scores (5-15 scale)."""
-
-    EXCELLENT = "excellent", "excellent"
-    GOOD = "good", "good"
-    POOR = "poor", "poor"
 
 
 EMAIL_TEMPLATE_PROMPT_FILES = {
