+++ src/research_ai/tests/test_constants.py
@@ -1,12 +1,6 @@
 from django.test import TestCase
 
-from research_ai.constants import (
-    ExpertiseLevel,
-    Gender,
-    OverallRating,
-    Region,
-    ReviewStatus,
-)
+from research_ai.constants import ExpertiseLevel, Gender, Region
 
 
 class ConstantsTests(TestCase):
@@ -21,11 +15,3 @@ class ConstantsTests(TestCase):
     def test_gender_choices(self):
         self.assertEqual(Gender.MALE, "male")
         self.assertEqual(Gender.ALL_GENDERS, "all_genders")
-
-    def test_review_status_choices(self):
-        self.assertEqual(ReviewStatus.PENDING, "pending")
-        self.assertEqual(ReviewStatus.COMPLETED, "completed")
-
-    def test_overall_rating_choices(self):
-        self.assertEqual(OverallRating.EXCELLENT, "excellent")
-        self.assertEqual(OverallRating.POOR, "poor")
