+++ src/researchhub/urls.py
@@ -276,6 +276,7 @@ urlpatterns = [
     path("api/permissions/", researchhub.views.permissions, name="permissions"),
     path("api/search/", include(search.urls)),
     path("api/research_ai/", include("research_ai.urls")),
+    path("api/ai_peer_review/", include("ai_peer_review.urls")),
     # Referral endpoints
     path("api/referral/", include("referral.urls")),
     # Organization endpoints
